Trash page populating
Trash page populating. Got the trash folder linked to the trash page.
It can look at the trash folder query the database based on the files
in the folder. If the files in the folder are found in the database
then they will show up… But we should probably make a status on the
schema for if the file is “indexed” “unindexed” or “trashed”